Add 1d2 and 1d3 support and more fixes
Add support to 1d2 and 1d3 rolls. In attack rolls it shows the rolls of 1d4 and 1d6 animation respectively and aply a formula of round(dice/2)

In macros 1d2 and 1d3 rolls are not shown but send to chat directly as messages.

Some rework on macros

Inventory improvements and adding carried capacity based on Strength and coin weight added (0.02 lbs)